Removed now unused VariantEditorFactory and VariantPropertyManager
These have been replaced by the new property framework.
