+++ tiled.qbs
@@ -29,7 +29,6 @@ Project {
         "src/libtiled",
         "src/libtiledquick",
         "src/plugins",
-        "src/qtpropertybrowser",
         "src/qtsingleapplication",
         "src/terraingenerator",
         "src/tiled/libtilededitor.qbs",
